feat: show username in index
